+++ src/features/post/__tests__/index.ts
@@ -2,4 +2,4 @@ export { PostServiceMocks } from "./mocks/post-service.mock";
 
 export type { MockPostService } from "./mocks/post-service.mock";
 
-export { mockPostDetailData } from "./fixtures";
+export { mockPostDetailData, mockPostsData } from "./fixtures";
