+++ src/shared/libs/index.ts
@@ -1,0 +1,2 @@
+export { formatDate, getRandomDate } from "./date";
+export { BaseError } from "./errors";
